分拆views视图
