+++ fastrunner/views/api.py
@@ -1,0 +1,120 @@
+from django.core.exceptions import ObjectDoesNotExist
+from rest_framework.viewsets import GenericViewSet
+from fastrunner import models, serializers
+from rest_framework.response import Response
+from fastrunner.utils import response
+from fastrunner.utils.parser import Format, Parse
+from django.db import DataError
+
+
+class APITemplateView(GenericViewSet):
+    """
+    API操作视图
+    """
+    serializer_class = serializers.APISerializer
+    queryset = models.API.objects
+    """使用默认分页器"""
+
+    def list(self, request):
+        """
+        接口列表 {
+            project: int,
+            node: int
+        }
+        """
+
+        node = request.query_params["node"]
+        project = request.query_params["project"]
+
+        queryset = self.get_queryset().filter(project__id=project, relation=node).order_by('-update_time')
+        pagination_queryset = self.paginate_queryset(queryset)
+        serializer = self.get_serializer(pagination_queryset, many=True)
+
+        return self.get_paginated_response(serializer.data)
+
+    def add(self, request):
+        """
+        新增一个接口
+        """
+
+        api = Format(request.data)
+        api.parse()
+
+        api_body = {
+            'name': api.name,
+            'body': api.testcase,
+            'url': api.url,
+            'method': api.method,
+            'project': models.Project.objects.get(id=api.project),
+            'relation': api.relation
+        }
+
+        try:
+            models.API.objects.create(**api_body)
+        except DataError:
+            return Response(response.DATA_TO_LONG)
+
+        return Response(response.API_ADD_SUCCESS)
+
+    def update(self, request, **kwargs):
+        """
+        更新接口
+        """
+        pk = kwargs['pk']
+        api = Format(request.data)
+        api.parse()
+
+        api_body = {
+            'name': api.name,
+            'body': api.testcase,
+            'url': api.url,
+            'method': api.method,
+        }
+
+        try:
+            models.API.objects.filter(id=pk).update(**api_body)
+        except ObjectDoesNotExist:
+            return Response(response.API_NOT_FOUND)
+
+        return Response(response.API_UPDATE_SUCCESS)
+
+    def delete(self, request, **kwargs):
+        """
+        删除一个接口 pk
+        删除多个
+        [{
+            id:int
+        }]
+        """
+
+        try:
+            if kwargs.get('pk'):  # 单个删除
+                models.API.objects.get(id=kwargs['pk']).delete()
+            else:
+                for content in request.data:
+                    models.API.objects.get(id=content['id']).delete()
+
+        except ObjectDoesNotExist:
+            return Response(response.API_NOT_FOUND)
+
+        return Response(response.API_DEL_SUCCESS)
+
+    def single(self, request, **kwargs):
+        """
+        查询单个api，返回body信息
+        """
+        try:
+            api = models.API.objects.get(id=kwargs['pk'])
+        except ObjectDoesNotExist:
+            return Response(response.API_NOT_FOUND)
+
+        parse = Parse(eval(api.body))
+        parse.parse_http()
+
+        resp = {
+            'id': api.id,
+            'body': parse.testcase,
+            'success': True,
+        }
+
+        return Response(resp)
