+++ uwsgi.ini
@@ -1,0 +1,20 @@
+# myweb_uwsgi.ini file
+[uwsgi]
+
+# Django-related settings
+
+project = FasterRunner
+base = /opt
+
+
+chdir = %(base)/%(project)
+module = %(project).wsgi:application
+
+
+master = true
+processes = 5
+
+
+socket = %(base)/%(project)/%(project).sock
+chmod-socket = 666
+vacuum = true
